mise à jour du programme pour un code plus clair
